add pipeline_writer attr for pipelines to customize their output
