+++ fpr/run_pipeline.py
@@ -10,7 +10,6 @@ import functools
 import logging
 import os
 import sys
-from typing import IO
 import json
 
 import rx
@@ -70,13 +69,6 @@ def parse_args():
     return parser.parse_args()
 
 
-def on_next_save_to_jsonl(outfile: IO, item):
-    log.debug("saving final pipeline item to {0}:\n{1}".format(outfile, item))
-    line = "{}\n".format(json.dumps(item))
-    outfile.write(line)
-    log.debug("wrote jsonl to {0}:\n{1}".format(outfile, line))
-
-
 def on_serialize_error(pipeline_name, e, *args):
     log.error(
         "error serializing result for {} pipeline:\n{}".format(
@@ -127,7 +119,7 @@ def main():
             )
         ),
     ).subscribe(
-        on_next=functools.partial(on_next_save_to_jsonl, args.outfile),
+        on_next=functools.partial(getattr(pipeline, "writer"), args.outfile),
         on_error=functools.partial(on_error, args.pipeline_name),
         on_completed=functools.partial(on_completed, loop=loop),
         scheduler=aio_scheduler,
